refactor: move View Plan button test to plan_mode.spec.ts and verify plan content before collapsing panel
